requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-socketio==5.3.5
flask-cors==4.0.0
flask-login==0.6.3
bcrypt==4.1.2
python-socketio==5.10.0
eventlet==0.33.3
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
sentence-transformers==2.2.2
openpyxl==3.1.2
